Refactor UX into dedicated components
